Initial Commit
In this commit, I set up the project structure and wrote the tutorial
README's to go with it. Each directory in the project root directory is
it's own CMake project.

The next step is to build example applications for parts 1, 2 and 3 to
demonstrate the concepts I discuss in the README's.